third_party/WebKit/LayoutTests/svg/custom/polyline-invalid-points-expected.txt

CONSOLE ERROR: line 4: Error: <polyline> attribute points: Expected number, "…100, 100 0, 100 INF INF".
